Update tests. Remove unused classes
