+++ app/src/main/AndroidManifest.xml
@@ -57,12 +57,6 @@
             android:windowSoftInputMode="adjustResize" />
 
         <activity
-            android:name=".activities.RegisterActivity"
-            android:parentActivityName=".activities.LoginActivity"
-            android:windowSoftInputMode="adjustResize">
-
-        </activity>
-        <activity
             android:name=".activities.NavigationActivity"
             android:launchMode="singleTop"
             android:screenOrientation="portrait">
