+++ app/src/main/java/com/testableapp/models/PlacesModel.java
@@ -24,15 +24,6 @@ public class PlacesModel {
         return INSTANCE;
     }
 
-    public Observable<ApiResponse<List<Place>>> getMockedPlaces(@NonNull final String criteria) {
-        return Observable.range(0, 5).map(integer ->
-                new Place("Genova 202" + integer.intValue(),
-                        new LatLng(-38.0635712197085, -57.55051396970849)))
-                .filter(place -> place.formattedAddress.contains(criteria))
-                .toList().map(places -> new ApiResponse.Builder<List<Place>>()
-                        .withData(places).withStatus("OK").build()).toObservable();
-    }
-
     public Observable<ApiResponse<List<Place>>> getPlaces(@NonNull final String criteria) {
         return getService().getPlaces(criteria);
     }
